requirements.txt
streamlit
supabase
pandas
python-dotenv
plotly
openpyxl
scipy
numpy
